com.unity.test-framework.performance@2.8.0-preview ## [2.8.0] - 2021-03-16
Add overloads to measurements for overriding sample unit
Fix cases where cleanup throws an exception

## [2.7.0] - 2021-02-19

Reduce metadata overhead when running locally by caching dependencies
Remove the need for link.xml
Restructured documentation
Fixed method measurement IterationsPerMeasurement

+++ Editor/TestRunBuilder.cs
@@ -76,16 +76,14 @@ namespace Unity.PerformanceTesting.Editor
 
         public void Cleanup()
         {
-            if (File.Exists(Utils.TestRunPath))
-            {
-                File.Delete(Utils.TestRunPath);
-                File.Delete(Utils.TestRunPath + ".meta");
-            }
+            if (File.Exists(Utils.TestRunPath)) { File.Delete(Utils.TestRunPath); }
+            const string metaPath = Utils.TestRunPath + ".meta";
+            if (File.Exists(metaPath)) {File.Delete(metaPath);}
 
             if (EditorPrefs.GetBool(cleanResources) && Directory.Exists("Assets/Resources"))
             {
                 Directory.Delete("Assets/Resources/", true);
-                File.Delete("Assets/Resources.meta");
+                if(File.Exists("Assets/Resources.meta")) {File.Delete("Assets/Resources.meta");}
             }
 
             AssetDatabase.Refresh();
